fix: use floorf() instead of std::floor() because other parts of the codebase also use floorf()

+++ libraries/AC_DroneShowManager/AC_DroneShowManager_Telemetry.cpp
@@ -203,7 +203,7 @@ uint8_t* AC_DroneShowManager::_fill_drone_show_status_packet_buffer(uint8_t* buf
     } else if (elapsed_time_sec <= -32768) {
         encoded_elapsed_time = -32768;
     } else {
-        encoded_elapsed_time = static_cast<int16_t>(std::floor(elapsed_time_sec));
+        encoded_elapsed_time = static_cast<int16_t>(floorf(elapsed_time_sec));
     }
 
     /* fill the packet. Note that in the first four bytes we _always_ put the
